Show N26 XP ladder on the stats table and track Available advancements Surface rank progress and a touch-friendly ladder tooltip on XP cells, and replace the Advancements header XP progress with an Available count for cumulative-XP editions.
